This update mainly for utilizing PyInstall to create a runnable .exe file for the game to run on machines without python/pygame and their dependencies etc. PyInstall added the build, dist and .spec folders and file. Spec was used to add the data (images) build has all the PyInstall files used to build the .exe and dist contains the actual single executable file.
Changes to the code itself in the main python script Project_Eterna.py include adding a new function resource_path() to load images from os directory which should allow it to function correctly with any build/machine. Also, PyInstaller needed the "sys._MEIPASS" to store the path for the image files in resources/art

Unfortunately had to reimplement the wall of code for the health, energy and xp bars instead of the loop I had implemented to eliminate. This was due to having to use the resource_path() function to load the health bar images and assigning new variables to each. might be a way around this to look into for future.