More art updates. Added some details to the Amuna buildings (shop, inn and academia) added some texture overlays to certain parts of the seldon background.
Added facing directions for the npc and associated art. This also included a new update() method for the npc class so that the images could be updated when a new facing direction is chosen for them. It will pick an npc and direction at random at 180 ticks (pygame.time.get_ticks())