Added a new start screen.
Updated a lot of image files and URLs to match the new resolution

Fixed scaling with the new resolutions. Resolution can be selected at start screen and will apply to everything.

Started working on battle sequence with the new separate iteration. Has a good start but still needs a lot of work.