lots of updates to scaling and shop/inventory
most of it seems to be correct and working, as far as positions of objects, their images, shop buying and selling, etc. for each resolution mode.

Next thing to add back in will be the inn instance and adding the sleep function.

Finally, at a point with the scaling and aspect ratio and code re-write that I can get back to adding new things soon.