Cleaned up a lot of code.
Removed unnecessary lines and added new functions when multiple lines of code were repeated. So functions for drawing message text and updating player status were created.

Fixed issue where combat info was not showing in message box after combat scenario. this was due to the text updates applying to the combats specific message box. had to create a dictionary and return it to the main loop to update the message box there.

finished implementing the inn instance. player can now rest to fully recover health and hp. sleep screen and transition to and from was also implemented.

academia instance started. partially implemented. works but art and functionality are not finished.