Added player offensive and defensive stats. These stats are based on the players currently equipped gear. Applies to combat appropriately and is shown on character sheet.
Fixed bug where items where disappearing after being equipped and un-equipped multiple times in a row. This no longer happens and was due to chest item being removed with weapon and not checked properly before being removed, so sometimes it would remove an empty string and replace it with an empty string in the inventory list, which would cause the item to be lost.

Added a window for leveling up so that its more obvious to player when it happens, as the level up text in message box can be overridden sometimes.