Quest system has the beginnings of its implementation.
Npcs can now be interacted with and will start a new screen, similar to the enemy encounter. But will offer a button to see their available quest and player can accept or decline

If quest is accepted then the journal is updated and the condition for that quest is set to true.