Brought the sprite sheet class over to the main project file so that the pygame display would not have to be re-initialized in the resource_urls.py
This caused quite a few cascading changes including having to move back the bar update functions to avoid a circular import issue.

Everything seems to be working, just needs some more cleaning up and polish. And need to figure out how to get the transparency back for the image backgrounds.