Lots of updates.
Fixed many sprites from the sprite sheets by cleaning up lines, positions of sprites/images etc.

Added an animation for the hearth function and a new screen to go with it. Also, an actual hearth stone that is attuned by default, but in future updates when more are added they will be able to be changed.

Added better time measurements for enemy and npc movement. As well as the message box clear. Noticed a big improvement already from that.

Added hair to default amuna "stan" sprites that didn't previously have it.

More things I think that I can't remember. This was a big update tho. Lots of polish added with functions/sprites and code.