Discovered there was more happening in the event loops than previously thought, such as collision detections for certain objects happening in every event for loop and some other things. Re-wrote a lot of code to fix this, hoping it improves performance.
Added to Celeste's quest to find her dog Nede. Nede now exists in the Stardust outpost and can be interacted with to complete the quest. Code and associated artwork for sprite also added.

Fixed an issue where Garan would continue to give player weapon/gear after already given to a player on a save file, because the condition was not being saved/loaded and checked. Condition is now checked and this bug should not occur.