Finished stardust post, added the ability to upgrade player's baseline offense and defense stats. This is done by trading 4 stars (1 each gained by questing) for either option. Allows for more offense as a fighter, more defense as a mage, or either for a scout.
Added a system that highlights quest targets when player is on that quest. It does this by updating the object's sprite to a highlighted version when the player's quest status relating to that object is true.

Added a new pop-up completion for quests when quest is complete, showing the reward clearly and removing that text from the message box.